data_bases/ingestor/requirements.txt
psycopg2-binary>=2.9,<3.0
PyMySQL>=1.1,<2.0
pymongo>=4.7,<5.0